Add jump-to-top functionality
