+++ app/src/main/java/com/capyreader/app/ui/articles/ArticleScaffold.kt
@@ -1,16 +1,7 @@
 package com.capyreader.app.ui.articles
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.asPaddingValues
-import androidx.compose.foundation.layout.calculateEndPadding
-import androidx.compose.foundation.layout.calculateStartPadding
-import androidx.compose.foundation.layout.displayCutout
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DrawerState
 import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ModalDrawerSheet
@@ -24,12 +15,11 @@ import androidx.compose.material3.adaptive.navigation.ThreePaneScaffoldNavigator
 import androidx.compose.material3.adaptive.navigation.rememberListDetailPaneScaffoldNavigator
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
+import com.capyreader.app.ui.components.safeEdgePadding
 import com.capyreader.app.ui.theme.CapyTheme
 
 @OptIn(ExperimentalMaterial3AdaptiveApi::class)
@@ -51,7 +41,7 @@ fun ArticleScaffold(
         },
     ) {
         ListDetailPaneScaffold(
-            modifier = Modifier.safePadding(),
+            modifier = Modifier.safeEdgePadding(),
             directive = scaffoldNavigator.scaffoldDirective,
             value = scaffoldNavigator.scaffoldValue,
             listPane = {
@@ -66,21 +56,6 @@ fun ArticleScaffold(
             }
         )
     }
-}
-
-@Composable
-private fun Modifier.safePadding(): Modifier {
-    val layoutDirection = LocalLayoutDirection.current
-    val displayCutout = WindowInsets.displayCutout.asPaddingValues()
-    val startPadding = displayCutout.calculateStartPadding(layoutDirection)
-    val endPadding = displayCutout.calculateEndPadding(layoutDirection)
-
-    return padding(
-        PaddingValues(
-            start = startPadding,
-            end = endPadding
-        )
-    )
 }
 
 @OptIn(ExperimentalMaterial3AdaptiveApi::class)
