+++ app/src/main/java/com/capyreader/app/ui/articles/FeedList.kt
@@ -32,6 +32,7 @@ import com.jocmp.capy.ArticleStatus
 import com.jocmp.capy.Feed
 import com.jocmp.capy.Folder
 import com.capyreader.app.R
+import com.capyreader.app.ui.components.safeEdgePadding
 import com.capyreader.app.ui.fixtures.FeedPreviewFixture
 import com.capyreader.app.ui.fixtures.FolderPreviewFixture
 import com.capyreader.app.ui.navigationTitle
@@ -60,6 +61,7 @@ fun FeedList(
 
     Column(
         Modifier
+            .safeEdgePadding()
             .fillMaxSize()
     ) {
         Column(
