+++ app/src/main/java/com/capyreader/app/ui/articles/FilterAppBarTitle.kt
@@ -1,9 +1,18 @@
 package com.capyreader.app.ui.articles
 
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
@@ -17,32 +26,51 @@ fun FilterAppBarTitle(
     filter: ArticleFilter,
     allFeeds: List<Feed>,
     folders: List<Folder>,
+    onRequestJumpToTop: () -> Unit
 ) {
     val text = when (filter) {
         is ArticleFilter.Articles -> stringResource(filter.articleStatus.navigationTitle)
         is ArticleFilter.Feeds -> {
             allFeeds.find { it.id == filter.feedID }?.title
         }
+
         is ArticleFilter.Folders -> {
             folders.find { it.title == filter.folderTitle }?.title
         }
     }.orEmpty()
 
-    Text(
-        text,
-        maxLines = 1,
-        overflow = TextOverflow.Ellipsis
-    )
+    Box(
+        contentAlignment = Alignment.CenterStart,
+        modifier = Modifier
+            .fillMaxSize()
+            .clickable(
+                indication = null,
+                interactionSource = remember { MutableInteractionSource() }
+            ) {
+                onRequestJumpToTop()
+            }
+    ) {
+        Text(
+            text,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis
+        )
+    }
+
 }
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Preview
 @Composable
 fun FilterAppBarTitlePreview() {
     MaterialTheme {
-        FilterAppBarTitle(
-            filter = ArticleFilter.default(),
-            allFeeds = emptyList(),
-            folders = emptyList(),
-        )
+        TopAppBar(title = {
+            FilterAppBarTitle(
+                filter = ArticleFilter.default(),
+                allFeeds = emptyList(),
+                folders = emptyList(),
+                onRequestJumpToTop = {}
+            )
+        })
     }
 }
